requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
openai==1.30.1
# chromadb==0.5.0  # Windows'ta C++ build tools gerektirir, in-memory fallback kullanılıyor
hypothesis==6.100.0
pytest==8.2.0
pytest-asyncio==0.23.6
alembic==1.13.1
streamlit==1.35.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiohttp==3.9.5
